Export ClassroomListItem to another file
